Renamed demo file to SearchAppBar and updated import accordingly.
